refactor(captcha): rename `verification` to `captcha`
