+++ backend/internal/captcha/handler.go
@@ -1,4 +1,4 @@
-package verification
+package captcha
 
 import (
 	"errors"
