+++ backend/internal/captcha/handler_test.go
@@ -1,4 +1,4 @@
-package verification
+package captcha
 
 import (
 	"encoding/json"
