+++ backend/internal/captcha/routes.go
@@ -1,4 +1,4 @@
-package verification
+package captcha
 
 import (
 	"github.com/gin-gonic/gin"
