+++ backend/internal/captcha/service_test.go
@@ -1,4 +1,4 @@
-package verification
+package captcha
 
 import (
 	"context"
